+++ app/__init__.py
@@ -21,6 +21,13 @@ def create_app():
     db.init_app(app=app)
 
     from app.app import bp as main_bp
+    from app.routes.products import product_bp
+    from app.routes.machines import machine_bp
+    from app.routes.stock import stock_bp
+
     app.register_blueprint(main_bp)
+    app.register_blueprint(product_bp)
+    app.register_blueprint(machine_bp)
+    app.register_blueprint(stock_bp)
 
     return app
